requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
matplotlib
scikit-learn
tensorflow
comet-ml
dvc